+++ src/ctests/matrix-hl.c
@@ -109,10 +109,10 @@ int main(int argc, char **argv)
    if (event[0] == PAPI_FP_INS) {
       /*     Compare measured FLOPS to expected value */
       tmp = 2 * (long long) (NROWS1) * (long long) (NCOLS2) * (long long) (NCOLS1);
-      if (abs((int)values[0] - (int)tmp) > tmp * 0.05) {
+      if (abs((int)values[0] - (int)tmp) > (double)tmp * 0.05) {
          /*     Maybe we are counting FMAs? */
          tmp = tmp / 2;
-         if (abs((int)values[0] - (int)tmp) > tmp * 0.05) {
+         if (abs((int)values[0] - (int)tmp) > (double)tmp * 0.05) {
             printf("\n" TAB1, "Expected operation count: ", 2 * tmp);
             printf(TAB1, "Or possibly (using FMA):  ", tmp);
             printf(TAB1, "Instead I got:            ", values[0]);
