+++ src/ctests/tenth.c
@@ -204,9 +204,9 @@ int main(int argc, char **argv)
       tmax = (long long) (sum[1] + sum[2]);
 #endif
 
-      printf ("percent error: %f\n", (float)(abs((int)(tmax - sum[0]))*100)/sum[0]);
-      tmin = (long long) (tmax * 0.8);
-      tmax = (long long) (tmax * 1.2);
+      printf ("percent error: %f\n", (float)(abs((int)(tmax - sum[0]))*100/sum[0]));
+      tmin = (long long)((double)tmax * 0.8);
+      tmax = (long long)((double)tmax * 1.2);
       if (sum[0] > tmax || sum[0] < tmin)
          test_fail(__FILE__, __LINE__, CACHE_LEVEL, 1);
    }
